function/requirements.txt
azure-common==1.1.28
azure-core==1.25.1
azure-functions==1.11.2
azure-servicebus==7.8.0
certifi==2022.9.14
charset-normalizer==2.1.1
cycler==0.10.0
idna==3.4
isodate==0.6.1
kiwisolver==1.3.1
matplotlib==3.3.4
msrest==0.7.1
numpy==1.20.1
oauthlib==3.2.1
pandas==1.2.3
patsy==0.5.1
Pillow==8.1.2
psycopg2-binary==2.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.28.1
requests-oauthlib==1.3.1
scipy==1.6.1
seaborn==0.11.1
six==1.15.0
statsmodels==0.12.2
typing-extensions==4.3.0
uamqp==1.6.0
urllib3==1.26.12
